Broke out transitionToTracklistForm function to be imported into main.js. #2
